Added   LAMMPS can be used in CrySPY

+++ CrySPY/interface/QE/calc_files_qe.py
@@ -33,4 +33,6 @@ def clean_calc_files_qe(work_path):
     # #        os.remove(work_path+f)
 
     # ---------- clear stat file
-    os.remove(work_path+'stat_job')
+    if os.path.isfile(work_path+'stat_job'):
+        os.remove(work_path+'stat_job')
+
